requirements.txt
azure-identity==1.16.0
openai==1.23.6
python-dotenv==1.0.1